+++ Kernel/Modules/AgentTicketBulk.pm
@@ -145,6 +145,7 @@ sub Run {
                 Data         => \%TypeList,
                 SelectedID   => $GetParam{TypeID},
                 PossibleNone => 1,
+                Translation  => 0,
             };
         }